도메인쪽에서 테스트에서만 쓰이는 생성자 제거
